locations/views.py
from rest_framework import viewsets
from locations.models import Country, State
from locations.serializers import CountrySerializer, StateSerializer


class StateViewSet(viewsets.ModelViewSet):
    queryset = State.objects.all()
    serializer_class = StateSerializer


class CountryViewSet(viewsets.ModelViewSet):
    queryset = Country.objects.all()
    serializer_class = CountrySerializer
