requirements.txt
appdirs==1.4.3
astroid==2.2.5
attrs==19.1.0
black==19.3b0
boto3==1.9.183
botocore==1.12.183
certifi==2019.6.16
chardet==3.0.4
Click==7.0
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.3
django-extensions==2.1.9
django-storages==1.7.1
djangorestframework==3.9.4
docutils==0.14
drf-yasg==1.16.0
idna==2.8
inflection==0.3.1
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==6.1.0
pyasn1==0.4.5
pylint==2.3.1
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
requests==2.22.0
rsa==3.4.2
ruamel.yaml==0.15.97
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.3
wrapt==1.11.2
